refactor: move all packages to internal since none of them are intended for external use
