+++ internal/functioncheck/README.md
