+++ internal/functioncheck/alloweditems.go
