+++ internal/git/git.go
@@ -2,23 +2,7 @@ package git
 
 import (
 	"fmt"
-	"os"
 )
-
-// Checks if needed enviorment variables are present in the .env file
-func CheckRequiredEnvironmentVariables() error {
-	vars := map[string]string{
-		"GITHUB_USER":         os.Getenv("GITHUB_USER"),
-		"GITHUB_TOKEN":        os.Getenv("GITHUB_TOKEN"),
-		"GITHUB_ORGANISATION": os.Getenv("GITHUB_ORGANISATION"),
-	}
-	for key, value := range vars {
-		if value == "" {
-			return fmt.Errorf("%s environment variable not set", key)
-		}
-	}
-	return nil
-}
 
 // Clone or open the repo & pull the latest changes into targetDirectory
 func Get(repoURL string, targetDirectory string) error {
